+++ StevenDimDoors/mod_pocketDim/items/ItemStableFabric.java
@@ -2,12 +2,12 @@ package StevenDimDoors.mod_pocketDim.items;
 
 import java.util.List;
 
-import StevenDimDoors.mod_pocketDim.CommonTickHandler;
 import StevenDimDoors.mod_pocketDim.LinkData;
 import StevenDimDoors.mod_pocketDim.SchematicLoader;
 import StevenDimDoors.mod_pocketDim.Spells;
 import StevenDimDoors.mod_pocketDim.mod_pocketDim;
 import StevenDimDoors.mod_pocketDim.helpers.dimHelper;
+import StevenDimDoors.mod_pocketDim.ticking.CommonTickHandler;
 import StevenDimDoors.mod_pocketDimClient.ClientTickHandler;
 
 import net.minecraft.block.Block;
